+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - feats for nimble characters
 - leveling for nimble characters
 - rolls for nimble characters
+- rest options for nimble characters
 
 ## [0.5.2] - 2026-07-24
 ### Added